Support Git tags
One of the dependencies in
[raylib](https://github.com/raysan5/raylib/blob/master/build.zig.zon)
has git rev 3.1.50, causing Nix to complain it is not a valid sha1 hash.
I added a naive check that checks if the rev is exactly 40 characters.
The only time I see this becoming an issue is if a tag's length is
exactly 40 characters, but that happening in the wild is very unlikely.