Move URL parsing and corner case handling to the Dependency struct
Prior to this change, URL parsing was scattered across `parse.zig` and `Dependency.zig`.
However, I've come to realize that it best belongs to the `Dependency` itself.

This change moves the parsing logic to the `Dependency` constructor. It also introduces
redirect following. I've noticed that `nix flake prefetch` resolves redirects, so the URLs
in the dependency map ended up being both the unresolved and resolved versions, leading to
mismatches. So the URL redirects are completely resolved before the Dependency is stored
in the dependency map.