Update used Zig version to a newer 0.16 nightly
Prior to this change, we had used zigpkgs.master-2025-11-13. This version
has a bug that occured for us in the fn `resolveRedirects`. This bug is
[#25811](https://github.com/ziglang/zig/issues/25811), which has been solved in
[#25909](https://github.com/ziglang/zig/pull/25909).

This change updates the used Zig version to `zigpkgs.master-2026-02-03` (which uses
`0.16.0-dev.2471+e9eadee00`) in order to use a Zig version with these bugs fixed.
That fixed the bug for us too.