In case a dependency is already known, don't override the existing one.
Prior to this change, a previously-known dependency was simply overridden. This caused a memory leak
as the new hash and url was only freed once in the `deps.deinit()`.

This change prevents the override and frees the memory right away. It also includes a safety check
to see if the previously-known `Dependency` has the same URL too.