Factor out URL generation for `nix flake prefetch` and fix it
Prior to this change, the URL to be used by `nix flake prefetch` was computed
"inline". It also contained a bug in case the dependency has both a `rev` and
and its URL contained `?ref=...`.

This change factors it out to the method `Dependency.flakePrefetchRef`. It also
contains tests and a fix. Without the fix, the 3rd testcase wouldn't work.