Add `overlay_0_16` to outputs and use it in CI for `run executable on zls`
Prior to this change, this CI step was run on `default_0_13`. This does not work with our
build system and code changes of a previous commit any more.

This change adds the `overlay_0_16` and uses it for the CI job. It also removes the linkage
flags from package.nix. Without this, we get the error
```
...
error: error: dynamic linking with lld requires at least one shared library
```