Don't store Dependency pointers in Workers, identify them by their key
Prior to this change, the `Worker` had a `dep: *Dependency` to identify the Dependency the worker is related to.
This dependency was later modified (nix_hash added to it). However, this is a pointer to a `StringHashMap`'s value
that is later being modified, further dependencies are added to the `deps` mapping as they are discovered.
This caused the pointer to become invalid, causing segmentation faults.

This change replaces the `dep: *Dependency` field in `Worker` with a `hash: []const u8` which is simply the
(owned copy of the) key identifying the dependency. Later on, dependencies are retrieved from the `deps` mapping
by their key, preventing the invalid pointer access described in the paragraph above.